move address.go to core/utils
